+++ src/main/java/group5/sebm/User/controller/dto/RegisterDto.java
@@ -1,4 +1,4 @@
-package group5.sebm.controller.dto;
+package group5.sebm.User.controller.dto;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
@@ -11,11 +11,19 @@ import lombok.Setter;
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
-public class LoginDto {
+public class RegisterDto {
   @NotBlank(message = "用户名不能为空")
   private String username;
 
   @NotBlank(message = "密码不能为空")
   @Size(min = 6, message = "密码至少6位")
   private String password;
+
+  @NotBlank(message = "请确认密码")
+  @Size(min = 6, message = "确认密码至少6位")
+  private String checkPassword;
+
+  //only phone number
+  @NotBlank(message = "电话号码不能为空")
+  private String phone;
 }
