+++ src/main/java/group5/sebm/User/dao/UserMapper.java
@@ -1,7 +1,7 @@
-package group5.sebm.dao;
+package group5.sebm.User.dao;
 
 import com.baomidou.mybatisplus.core.mapper.BaseMapper;
-import group5.sebm.entity.UserPo;
+import group5.sebm.User.entity.UserPo;
 
 /**
  * @author AL
