+++ src/main/java/group5/sebm/User/service/bo/Borrower.java
@@ -1,11 +1,9 @@
-package group5.sebm.service.bo;
+package group5.sebm.User.service.bo;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-
-import java.util.Date;
 
 /**
  * @description: Borrower class representing a user who borrows items.
