+++ src/main/resources/mapper/UserMapper.xml
@@ -2,9 +2,9 @@
 <!DOCTYPE mapper
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="group5.sebm.dao.UserMapper">
+<mapper namespace="group5.sebm.User.dao.UserMapper">
 
-    <resultMap id="BaseResultMap" type="group5.sebm.entity.UserPo">
+    <resultMap id="BaseResultMap" type="group5.sebm.User.entity.UserPo">
             <id property="id" column="id" jdbcType="BIGINT"/>
             <result property="username" column="username" jdbcType="VARCHAR"/>
             <result property="password" column="password" jdbcType="VARCHAR"/>
